# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top Maracana API",
        "description": "Documentation de l'API Top Maracana",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/competitions": {
            "get": {
                "tags": [
                    "Compétitions"
                ],
                "summary": "Lister toutes les compétitions publiques",
                "operationId": "5745adf0e6a72bffacfb3bc5ea52d7e4",
                "responses": {
                    "200": {
                        "description": "Liste des compétitions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "competitions": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tournoi Maracana"
                                                            },
                                                            "location": {
                                                                "type": "string",
                                                                "example": "Ouagadougou"
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "registration_open"
                                                            },
                                                            "start_date": {
                                                                "type": "string",
                                                                "example": "2026-03-15"
                                                            },
                                                            "end_date": {
                                                                "type": "string",
                                                                "example": "2026-03-25"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Compétitions"
                ],
                "summary": "Créer une nouvelle compétition",
                "operationId": "2c7a8ef1b623d697b29d901341e6a171",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "start_date",
                                    "end_date",
                                    "max_teams",
                                    "players_per_team"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tournoi Maracana"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Ouagadougou"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-25"
                                    },
                                    "max_teams": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "players_per_team": {
                                        "type": "integer",
                                        "example": 11
                                    },
                                    "registration_fee": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "prize_description": {
                                        "type": "string",
                                        "example": "Trophée + 500 000 FCFA"
                                    },
                                    "age_min": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "age_max": {
                                        "type": "integer",
                                        "example": 40
                                    },
                                    "poster_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compétition créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "competition": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tournoi Maracana"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "registration_open"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/competitions/my": {
            "get": {
                "tags": [
                    "Compétitions"
                ],
                "summary": "Lister les compétitions de l'organisateur connecté",
                "operationId": "b2db4c1a55c6d08a06f5cd79ffdfa2a9",
                "responses": {
                    "200": {
                        "description": "Mes compétitions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "competitions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Tournoi Maracana"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "registration_open"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/competitions/{id}/groups": {
            "get": {
                "tags": [
                    "Compétitions"
                ],
                "summary": "Groupes d'une compétition",
                "operationId": "3630e750c02dec5f2add0f0fb2975f6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groupes de la compétition"
                    },
                    "404": {
                        "description": "Compétition introuvable"
                    }
                }
            }
        },
        "/api/competitions/{id}/matches": {
            "get": {
                "tags": [
                    "Compétitions"
                ],
                "summary": "Matchs d'une compétition",
                "operationId": "be12bebc719b793560f52519a1973d64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matchs de la compétition"
                    },
                    "404": {
                        "description": "Compétition introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}